Removed Vagrant SSH password and Vagrant insecure SSH key pair
Vagrant user authentication an SSH key pair. A key pair is generated for
each operating system.

The Vagrant user account is locked out. SSH is the only means of access.

Encrypted the root password stored in Kickstart and Preseed
configuration files. The plain text root password is saved in the
root_user_password file. It is generated for each operating system.

Preseed and Kickstart configuration files are generated using the
kickstart script.